Phase 1: fix parse error, parallelize pipeline, optimize prompts
Server:
- Fix EmotionResult.Intensity type: *int → *float64 (fixes unmarshal error
  when DeepSeek returns float like 0.5)
- Fix Anthropic schema: emotion intensity "integer" → "number"
- Add thinking config: explicitly disable DeepSeek thinking mode
  via thinking.type="disabled" for both analyze and reflection

iOS:
- Parallelize pipeline: fire reflection API call (async let) while
  graph/arc processing runs, merge results after both complete
  (~15s → ~8s end-to-end)
- Limit known entities to 20 most recent (prompt size reduction)
- Truncate artifact textContent to 500 chars in API payload
- Replace 4s polling loop with NotificationCenter-based refresh
  (pipelineDidComplete notification posted on success/failure)
- Fix AnalyzeResponseMapper emotion intensity: Int? → Double?

+++ mory/mory/Infrastructure/Analysis/AnalyzeResponseMapper.swift
@@ -133,7 +133,7 @@ struct AnalyzeResponseMapper {
 struct AnalyzeResponseEnvelope: Codable, Sendable {
     struct Emotion: Codable, Sendable {
         var label: String
-        var intensity: Int?
+        var intensity: Double?
         var confidence: Double?
         var interpretation: String?
     }
